cursor_mcp/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
PyGithub==1.59.1
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
aiofiles==23.2.1
httpx>=0.27.0,<0.28.0
pytest==7.4.3 